requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1
django-cors-headers==4.4.0
django-grappelli==4.0.1
django-jazzmin==3.0.0
django-material-admin==1.8.6
django-redis==5.4.0
django-suit==0.2.28
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==23.0.0
inflection==0.5.1
invoke==2.2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
openapi==2.0.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
rich==13.8.0
rpds-py==0.20.0
setuptools==73.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
